+++ dotNetInstallerLib/ProcessComponent.cpp
@@ -44,31 +44,31 @@ void ProcessComponent::Wait(DWORD /* tt */)
 		L"WaitForSingleObject");
 }
 
-void ProcessComponent::ExecCmd(std::wstring command, DVLib::CommandExecutionMethod executionMethod, bool disableWow64FsRedirection)
+void ProcessComponent::ExecCmd(std::wstring command, DVLib::CommandExecutionMethod executionMethod, bool disableWow64FsRedirection, bool hideWindow)
 {
 	if (disableWow64FsRedirection)
 	{
 		auto_any<Wow64NativeFS *, close_delete> wow64_native_fs(new Wow64NativeFS());
-		ExecCmdCore(command, executionMethod);
+		ExecCmdCore(command, executionMethod, hideWindow);
 	}
 	else
 	{
-		ExecCmdCore(command, executionMethod);
+		ExecCmdCore(command, executionMethod, hideWindow);
 	}
 }
 
-void ProcessComponent::ExecCmdCore(std::wstring command, DVLib::CommandExecutionMethod executionMethod)
+void ProcessComponent::ExecCmdCore(std::wstring command, DVLib::CommandExecutionMethod executionMethod, bool hideWindow)
 {
 	PROCESS_INFORMATION process_info;
 	switch (executionMethod)
 	{
 		case DVLib::CemCreateProcess:
-			DVLib::RunCmd(command, & process_info);
+			DVLib::RunCmd(command, & process_info, 0, hideWindow);
 			m_process_handle = process_info.hProcess;
 			::CloseHandle(process_info.hThread);
 			break;
 		case DVLib::CemShellExecute:
-			DVLib::ShellCmd(command, NULL, & m_process_handle, main_window);
+			DVLib::ShellCmd(command, NULL, & m_process_handle, main_window, hideWindow);
 			break;
 	}
 }